+++ lib/src/models/api_response.dart
@@ -27,7 +27,7 @@ class ApiResponse {
 
   ///Convert json to [ApiResponse]
   factory ApiResponse.fromJson(Map<String, dynamic> json) => ApiResponse(
-        body: json['body'] as Map<String, dynamic>,
+        body: json['body'] as dynamic,
         baseUrl: json['baseUrl'] as String,
         method: json['method'] as String,
         statusCode: json['statusCode'] as int,
@@ -99,7 +99,7 @@ class ApiResponse {
   final Map<String, dynamic> request;
 
   ///Response data
-  final Map<String, dynamic> body;
+  final dynamic body;
 
   ///Request data type
   final String? contentType;
@@ -165,7 +165,7 @@ class ApiResponse {
     double? responseSize,
     Map<String, dynamic>? request,
     String? response,
-    Map<String, dynamic>? body,
+    dynamic body,
     String? contentType,
     String? headers,
     int? sendTimeout,
@@ -207,8 +207,8 @@ Base URL: $baseUrl
 Path: $path
 Method: $method
 Status Code: $statusCode
-Request Time: ${requestTime.toString()}
-Response Time: ${responseTime.toString()}
+Request Time: $requestTime
+Response Time: $responseTime
 Headers: $headers
 Query Params: $queryParameters
 Content Type: $contentType