+++ lib/src/view/api_detail_page.dart
@@ -50,43 +50,45 @@ class _ApiDetailsPageState extends State<ApiDetailsPage> {
             ),
           ],
         ),
-        body: DefaultTabController(
-          length: 3,
-          child: Column(
-            children: [
-              Material(
-                color: primaryColor,
-                child: TabBar(
-                  labelColor: Colors.white,
-                  unselectedLabelColor: Colors.white.withOpacity(0.8),
-                  tabs: [
-                    Tab(text: Localization.strings['overview']),
-                    Tab(text: Localization.strings['request']),
-                    Tab(text: Localization.strings['response']),
-                  ],
+        body: SafeArea(
+          child: DefaultTabController(
+            length: 3,
+            child: Column(
+              children: [
+                Material(
+                  color: primaryColor,
+                  child: TabBar(
+                    labelColor: Colors.white,
+                    unselectedLabelColor: Colors.white.withOpacity(0.8),
+                    tabs: [
+                      Tab(text: Localization.strings['overview']),
+                      Tab(text: Localization.strings['request']),
+                      Tab(text: Localization.strings['response']),
+                    ],
+                  ),
                 ),
-              ),
-              Expanded(
-                child: TabBarView(
-                  key: const Key('api_detail_tabbar_view'),
-                  children: [
-                    OverviewTabView(api: widget.api),
-                    _RequestTab(
-                      jsonPreviewType: _jsonRequestPreviewType,
-                      onShufflePreview: _shuffleRequestPreviewType,
-                      json: widget.api.request,
-                      prettyJson: widget.api.prettyJsonRequest,
-                    ),
-                    _ResponseTab(
-                      jsonPreviewType: _jsonResponsePreviewType,
-                      onShufflePreview: _shuffleResponsePreviewType,
-                      json: widget.api.body,
-                      prettyJson: widget.api.prettyJson,
-                    )
-                  ],
+                Expanded(
+                  child: TabBarView(
+                    key: const Key('api_detail_tabbar_view'),
+                    children: [
+                      OverviewTabView(api: widget.api),
+                      _RequestTab(
+                        jsonPreviewType: _jsonRequestPreviewType,
+                        onShufflePreview: _shuffleRequestPreviewType,
+                        json: widget.api.request,
+                        prettyJson: widget.api.prettyJsonRequest,
+                      ),
+                      _ResponseTab(
+                        jsonPreviewType: _jsonResponsePreviewType,
+                        onShufflePreview: _shuffleResponsePreviewType,
+                        json: widget.api.body,
+                        prettyJson: widget.api.prettyJson,
+                      )
+                    ],
+                  ),
                 ),
-              ),
-            ],
+              ],
+            ),
           ),
         ),
       ),
@@ -141,7 +143,7 @@ class _PreviewModeControl extends StatelessWidget {
             children: [
               Text(
                 Localization.strings['jsonPreviewMode']!,
-                style: context.textTheme.bodyText2!
+                style: context.textTheme.bodyMedium!
                     .toBold()
                     .withColor(primaryColor),
               ),
@@ -149,7 +151,7 @@ class _PreviewModeControl extends StatelessWidget {
                 onPressed: onPreviewPressed,
                 height: 34,
                 text: type,
-                style: context.textTheme.bodyText2!.toBold(),
+                style: context.textTheme.bodyMedium!.toBold(),
               )
             ],
           ),
@@ -233,7 +235,7 @@ class _ResponseTab extends StatelessWidget {
           width: double.maxFinite,
           child: SelectableText(
             prettyJson,
-            style: context.textTheme.bodyText1,
+            style: context.textTheme.bodyLarge,
             textDirection: TextDirection.ltr,
           ),
         );
@@ -301,7 +303,7 @@ class _RequestTab extends StatelessWidget {
           width: double.maxFinite,
           child: SelectableText(
             prettyJson,
-            style: context.textTheme.bodyText1,
+            style: context.textTheme.bodyLarge,
             textDirection: TextDirection.ltr,
           ),
         );