+++ lib/src/view/widgets/http_methods_menu.dart
@@ -42,7 +42,7 @@ class _HttpMethodsMenuState extends State<HttpMethodsMenu> {
               Chip(
                 label: Text(
                   _getMethodName(),
-                  style: context.textTheme.bodyText1!.withColor(Colors.white),
+                  style: context.textTheme.bodyLarge!.withColor(Colors.white),
                 ),
                 backgroundColor: methodColor(_getMethodName()),
               )
@@ -61,7 +61,7 @@ class _HttpMethodsMenuState extends State<HttpMethodsMenu> {
     );
   }
 
-  PopupMenuEntry _radioButton(
+  PopupMenuEntry<dynamic> _radioButton(
     String text,
     HttpMethod httpMethod,
   ) {