+++ test/src/view/api_detail_page_test.dart
@@ -10,9 +10,9 @@ import 'package:flutter_test/flutter_test.dart';
 import 'package:shared_preferences/shared_preferences.dart';
 
 void main() {
-  TestWidgetsFlutterBinding.ensureInitialized();
-  SharedPreferences.setMockInitialValues({});
-  final _sharedPreferencesManager = SharedPreferencesManager.getInstance();
+  final sharedPreferencesManager = SharedPreferencesManager.getInstance(
+    initData: false,
+  );
   testWidgets(
     'When page opened, three tabs should be loaded',
     (WidgetTester tester) async {
@@ -161,7 +161,7 @@ void main() {
         body: {'data': 'https://example.png'},
       );
 
-      await _sharedPreferencesManager.addApiResponse(api);
+      await sharedPreferencesManager.addApiResponse(api);
 
       await tester.pumpWidget(
         MaterialApp(
@@ -198,7 +198,7 @@ void main() {
         body: {'data': 'https://example.png'},
       );
 
-      await _sharedPreferencesManager.addApiResponse(api);
+      await sharedPreferencesManager.addApiResponse(api);
 
       await tester.pumpWidget(
         MaterialApp(